requirements.txt
goodwe
influxdb-client
schedule
requests
paho-mqtt
pymodbus==3.11.4
